feat: added recipe add/edit screen

+++ app/src/main/java/com/ihita/wholetthemcook/data/RecipeDao.kt
@@ -13,5 +13,5 @@ interface RecipeDao {
     suspend fun getAllRecipes(): List<Recipe>
 
     @Query("SELECT * FROM Recipe WHERE id = :recipeId")
-    suspend fun getRecipeById(recipeId: Int): Recipe
+    suspend fun getRecipeById(recipeId: Long): Recipe
 }
